Move/rename shc.log.generic to shc.data_logging and log adapters to shc.interfaces
