--- shc/log/__init__.py
@@ -1,2 +1,0 @@
-
-from .generic import AggregationMethod
