+++ test/interfaces/test_file_persistence.py
@@ -6,7 +6,7 @@ from pathlib import Path
 import aiofile.aio
 
 import shc.base
-from shc.log import file_persistence
+from shc.interfaces import file_persistence
 from .._helper import InterfaceThreadRunner
 from ..test_variables import ExampleTupleType
